# Improve performance by directly constructing AST from buffer