--- a/scripts/generate-ast-converters.js
+++ b/scripts/generate-ast-converters.js
@@ -3,3 +3,4 @@
 import './generate-rust-constants.js';
 import './generate-buffer-to-ast.js';
 import './generate-child-node-keys.js';
+import './generate-buffer-parsers.js';