--- a/scripts/generate-ast-converters.js
+++ b/scripts/generate-ast-converters.js
@@ -2,3 +2,4 @@
 
 import './generate-rust-constants.js';
 import './generate-buffer-to-ast.js';
+import './generate-keys.js';