--- a/test/function/samples/catch-rust-panic-parse/main.js
+++ b/test/function/samples/catch-rust-panic-parse/main.js
@@ -1,0 +1,1 @@
+assert.ok(true)