… CONTRIBUTING.md
… package.json
… scripts/find-config.js
… scripts/generate-buffer-to-ast.js
… scripts/generate-child-node-keys.js
… scripts/generate-keys.js
… scripts/generate-rust-constants.js
… scripts/perf-init.js
… scripts/perf.js
… scripts/tsconfig.json
… src/Module.ts
… src/ast/bufferParsers.ts
… src/ast/childNodeKeys.ts
… src/ast/keys.ts
… src/ast/nodes/ArrowFunctionExpression.ts
… src/ast/nodes/AssignmentExpression.ts
… src/ast/nodes/BlockStatement.ts
… src/ast/nodes/CallExpression.ts
… src/ast/nodes/CatchClause.ts
… src/ast/nodes/ClassBody.ts
… src/ast/nodes/ClassDeclaration.ts
… src/ast/nodes/DebuggerStatement.ts
… src/ast/nodes/ExportAllDeclaration.ts
… src/ast/nodes/ExportDefaultDeclaration.ts
… src/ast/nodes/ExportNamedDeclaration.ts
… src/ast/nodes/ExpressionStatement.ts
… src/ast/nodes/ForInStatement.ts
… src/ast/nodes/ForOfStatement.ts
… src/ast/nodes/FunctionDeclaration.ts
… src/ast/nodes/FunctionExpression.ts